Sort out the Scrooge/libthrift dependencies
* Move the dependencies to the `json` project where they belong
* Scrooge should be 4.5.0 to match what we use in other projects
* Downgrade libthrift to 0.9.1 to fix the incompatible equals() change

The reason for downgrading libthrift is that Scrooge is compiled against
libthrift 0.5.0-1 (Twitter's fork), which turns out to be incompatible
with 0.9.2 or newer. When it deserializes a struct containing one or
more optional fields that it doesn't understand, its `equals` method
delegates to the one provided by `libthrift`. But the signature of that
method changed between 0.9.1 and 0.9.2.

If we discover any more surprises like this, I recommend we downgrade
further to 0.5.0-1, which is known to work well with Scrooge.