Got Circe working for tags, sections and atoms
I think Content would work if the compile ever finished...

I've commented out the lines that take forever to compile. It looks like
compilation slowdown is (exponentionally?!) proportional to the number
of fields. We may have to write the biggest ones by hand, or maybe write
a macro specially for Thrift structs.