Replace Travis's macro with one of my own
We were using a macro to guide shapeless towards automatic derivation
of `Decoder` for Scrooge classes. This worked fine for some structs, but
compile time became absurdly long (infinite?) for complex structs such
as `Content`.

Replaced it with a new macro that skips the whole shapeless step and
explicitly generates a `Decoder` instance for a given struct type. This
simplifies things and makes compilation much faster.

Also added an `Encoder` to convert from json4s JValue to Circe Json
without having to go via a string.