Re-introduce automated Typescript NPM release
This is in addition to the Scala Maven release provided by the previous
commit. This change is only separated out from the previous commit to
show how the 'standard' release.yml has been modified for this dual
Scala & Typescript release workflow.

This is a first for gha-scala-library-release-workflow - so far that
workflow has only been used for publishing Scala projects to Maven - but
the earlier work done in PR #229 supported automated release to both Maven
and NPM (for NPM, we're talking Typescript artifacts generated from Thrift
by scrooge-extras), and so it was important to continue to that support.

Rather than teach the gha-scala-library-release-workflow reusable-release.yml
how to publish NPM artifacts, it was easier to just modify the standard
release.yml workflow that calls reusable-release.yml, adding an additional
NPM Release workflow job after the Maven release process.