Remove unnecessary duplication of tests
These tests were originally duplicated across two scala versions in
[PR #172](https://github.com/guardian/content-api-models/pull/172), when
new tests were added for scala 2.11. There’s not much info in the PR or
commit history, but it looks like the reason for duplicating the source
code was to use an older version of diffson-circe, where the API is
slightly different (v3 is the last released version for 2.11, so the v4
version in use in the tests at the time couldn’t be used in the 2.11
tests).

Since we now don’t support 2.11 anymore, the test source code doesn’t
need to differ across different scala versions. (On main at the moment
they differ slightly across the 2.12 and 2.11 tests, but only because
recent changes have been made to 2.12 and not 2.11.)

Since sbt runs tests that are not version-specific (e.g. those in
`test/scala` rather than `test/scala-2.x`) for every requested scala
version, there’s no need to duplicate the files when the tests are
intended to be the same.

As such, this commit makes the newer tests not version-specific anymore,
by moving them back to json/src/test/scala, deleting the second copy.