Try using circe-generic-extras to rename fields
To keep using the old names in the json produced and consumed by these
encoders and decoders, it should be possible to use circe-generic-extras
to define configured encoders and decoders that do the field renaming.

However, this currently fails because deriveConfiguredEncoder and
deriveConfiguredDecoder can’t tell that it’s possible to use fezziwig to
get the necessary shapeless representations they can normally get
already for case classes and traits.

It might be possible to update fezziwig to support using it with
circe-generic-extras (by outputting whatever deriveConfiguredEncoder
needs that deriveEncoder doesn’t), but it’s probably easier to avoid
using circe-generic-extras altogether: I’ll investigate that
next. (Keeping this commit around for history.)