Avoid using circe-generic-extras
Since the changes required are pretty straightforward, it’s easier to
use .prepare and .mapJson rather than get circe-generic-extras working
with fezziwig. However, this does mean we need to check this code quite
carefully: working directly with Json means we lose a lot of static
typing guarantees, which we will want to get back with tests.

(For example, as it is this doesn’t handle the case where only one of
the fields is present correctly!)